+++ spec/default_config_spec.rb
@@ -26,29 +26,29 @@ describe MAuth::Client do
 
     it 'guesses everything' do
       Dir.chdir('spec/config_root') do
-        MAuth::Client.default_config['app_uuid'].should == 'NORMAL-DEVELOPMENT'
+        expect(MAuth::Client.default_config['app_uuid']).to eq('NORMAL-DEVELOPMENT')
       end
     end
 
     it 'raises when it cannot find mauth config yml' do
-      expect { MAuth::Client.default_config }.to raise_error(MAuth::Client::ConfigurationError)
+      expect { MAuth::Client.default_config(:mauth_config_yml => "no_file_here") }.to raise_error(MAuth::Client::ConfigurationError)
     end
 
     it 'has root option specified' do
-      MAuth::Client.default_config(:root => 'spec/config_root')['app_uuid'].should == 'NORMAL-DEVELOPMENT'
+      expect(MAuth::Client.default_config(:root => 'spec/config_root')['app_uuid']).to eq('NORMAL-DEVELOPMENT')
     end
 
     it 'has root env variable specified' do
       %w(RAILS_ROOT RACK_ROOT APP_ROOT).each do |var|
         with_env(var => 'spec/config_root') do
-          MAuth::Client.default_config['app_uuid'].should == 'NORMAL-DEVELOPMENT'
+          expect(MAuth::Client.default_config['app_uuid']).to eq('NORMAL-DEVELOPMENT')
         end
       end
     end
 
     it 'has environment option specified' do
       Dir.chdir('spec/config_root') do
-        MAuth::Client.default_config(:environment => 'production')['app_uuid'].should == 'NORMAL-PRODUCTION'
+        expect(MAuth::Client.default_config(:environment => 'production')['app_uuid']).to eq('NORMAL-PRODUCTION')
       end
     end
 
@@ -56,7 +56,7 @@ describe MAuth::Client do
       Dir.chdir('spec/config_root') do
         %w(RAILS_ENV RACK_ENV).each do |var|
           with_env(var => 'production') do
-            MAuth::Client.default_config['app_uuid'].should == 'NORMAL-PRODUCTION'
+            expect(MAuth::Client.default_config['app_uuid']).to eq('NORMAL-PRODUCTION')
           end
         end
       end
@@ -65,25 +65,25 @@ describe MAuth::Client do
     it 'has Rails.environment specified' do
       Dir.chdir('spec/config_root') do
         with_rails(:environment => 'production') do
-          MAuth::Client.default_config['app_uuid'].should == 'NORMAL-PRODUCTION'
+          expect(MAuth::Client.default_config['app_uuid']).to eq('NORMAL-PRODUCTION')
         end
       end
     end
 
     it 'has mauth_config_yml option specified' do
-      MAuth::Client.default_config(:mauth_config_yml => 'spec/config_root/config/mauth.yml')['app_uuid'].should == 'NORMAL-DEVELOPMENT'
+      expect(MAuth::Client.default_config(:mauth_config_yml => 'spec/config_root/config/mauth.yml')['app_uuid']).to eq('NORMAL-DEVELOPMENT')
     end
 
     it 'has MAUTH_CONFIG_YML env var specified' do
       with_env('MAUTH_CONFIG_YML' => 'spec/config_root/config/mauth.yml') do
-        MAuth::Client.default_config['app_uuid'].should == 'NORMAL-DEVELOPMENT'
+        expect(MAuth::Client.default_config['app_uuid']).to eq('NORMAL-DEVELOPMENT')
       end
     end
 
     it 'has logger option specified' do
       Dir.chdir('spec/config_root') do
         logger = ::Logger.new(StringIO.new)
-        MAuth::Client.default_config(:logger => logger)['logger'].should == logger
+        expect(MAuth::Client.default_config(:logger => logger)['logger']).to eq(logger)
       end
     end
 
@@ -91,7 +91,7 @@ describe MAuth::Client do
       Dir.chdir('spec/config_root') do
         logger = ::Logger.new(StringIO.new)
         with_rails(:logger => logger) do
-          MAuth::Client.default_config['logger'].should == logger
+          expect(MAuth::Client.default_config['logger']).to eq(logger)
         end
       end
     end