+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 require 'timecop'
 require 'uuidtools'
 require 'json'
+require 'rack/mock'
 
 if (RUBY_VERSION.split('.').map(&:to_i) <=> [1, 9]) >= 0
   require 'simplecov'