+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # MAuth-Client History
+## v2.7.0
+- ability to pass custom headers into mauth-client and mauth-proxy
+- Upgraded to use newest version of Faraday Middleware
+- faraday_options now only get merged to the request (previously got merged into everything)
+
 ## v2.6.4
 
 - Less restrictive rack versioning to allow for more consumers.